// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const OfferStatus = {
  ACTIVE: 'ACTIVE',
  EXPIRED: 'EXPIRED',
  SOLD_OUT: 'SOLD_OUT',
  CANCELLED: 'CANCELLED'
} as const

export type OfferStatus = (typeof OfferStatus)[keyof typeof OfferStatus]


export const OrderStatus = {
  PENDING_PAYMENT: 'PENDING_PAYMENT',
  CONFIRMED: 'CONFIRMED',
  READY_FOR_PICKUP: 'READY_FOR_PICKUP',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
  NO_SHOW: 'NO_SHOW'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const PaymentMethod = {
  PIX: 'PIX',
  CREDIT_CARD: 'CREDIT_CARD'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const PaymentStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REFUSED: 'REFUSED',
  REFUNDED: 'REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const NotificationType = {
  ORDER_CONFIRMED: 'ORDER_CONFIRMED',
  PICKUP_REMINDER: 'PICKUP_REMINDER',
  ORDER_CANCELLED: 'ORDER_CANCELLED',
  REVIEW_REQUEST: 'REVIEW_REQUEST',
  NEW_ORDER: 'NEW_ORDER'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]